checked if map is rectangular and if the characters are valid

+++ lib/lib.h
@@ -6,7 +6,7 @@
 /*   By: joapedro <joapedro@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 12:38:06 by joapedro          #+#    #+#             */
-/*   Updated: 2025/07/09 11:27:25 by joapedro         ###   ########.fr       */
+/*   Updated: 2025/07/10 14:00:52 by joapedro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ char	*ft_strjoin(char *line, char *read_buffer);
 void	shift(char *read_buffer);
 char	*ft_strrchr(const char *s, int c);
 char	*ft_strchr(const char *s, int c);
-size_t	ft_strlen(const char *s);
+int	ft_strlen(const char *s);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 int		ft_strcmp(char *s1, char *s2);
 char	*ft_strstr(char *str, char *to_find);
